build: Add chromium prebuilt support to envsetup.sh && The core Makefile
This adds a chromium_prebuilt function to envsetup.sh that is invoked by lunch to check
whether the chromium prebuilts are up-to-date or not. If not, it will be built from source
and then the new source built version will be pulled during brunch/mka bacon to become the
new prebuilts for future builds.

This is all opt-in through the USE_PREBUILT_CHROMIUM flag. Without it being set to 1,
none of this would be ran, and regular operations will go on.

PS13:
-use export TARGET_DEVICE
-replace git -C with params compatible to old git versions

(vendor repository for script)

Change-Id: I8c010ddb1e52497cbb4874e33143e184c517d3c3

Conflicts:
	envsetup.sh